% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chol.R
\name{chol,AD_matrix-method}
\alias{chol,AD_matrix-method}
\title{overload of the built in (upper-triangular) Cholesky}
\usage{
\S4method{chol}{AD_matrix}(x, ...)
}
\description{
overload of the built in (upper-triangular) Cholesky
}
