% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{determinant,AD_matrix-method}
\alias{determinant,AD_matrix-method}
\title{Deterimant of AD_matrix type}
\usage{
\S4method{determinant}{AD_matrix}(x, logarithm = TRUE, ...)
}
\description{
Deterimant of AD_matrix type
}
