% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{fAD2-class}
\alias{fAD2-class}
\title{class representing a numeric vector, it's Jacobian matrix and an 3d array of Hessian matrices}
\description{
class representing a numeric vector, it's Jacobian matrix and an 3d array of Hessian matrices
}
