% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad.R
\name{independent}
\alias{independent}
\title{Set independent variable for first order derivatives}
\usage{
independent(x)
}
\arguments{
\item{x}{numeric vector to be differentiated with respect to}
}
\value{
a fAD vector with value being x and jacobian being the identity
}
\description{
Set independent variable for first order derivatives
}
\examples{
x <- independent(c(0,1))
y <- sum(square(x))/2
print(gradient(y)) # should be c(0,1)

# define some function
f <- function(x){ 
M <- matrix(x,2,2)
return(sum(solve(M,c(-1,2))))
}

# run function with regular variables
print(f(c(1,-1,2,4)))

# run function with AD-variables
x <- independent(c(1,-1,2,4))
y <- f(x)
print(value(y))
print(gradient(y))

}
