% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad.R
\name{independent2}
\alias{independent2}
\title{Set independent variable for first and second order derivatives}
\usage{
independent2(x)
}
\arguments{
\item{x}{numeric vector to be differentiated with respect to}
}
\value{
a fAD2 vector with value being x, jacobian being the identity, and hessian being zero
}
\description{
Set independent variable for first and second order derivatives
}
\examples{
x <- independent2(c(0,1))
y <- sum(square(x))/2
print(gradient(y)) # should be c(0,1)
print(hessian(y)) # should be identity
 
# define some function
f <- function(x){ 
M <- matrix(x,2,2)
return(sum(solve(M,c(-1,2))))
}

# run function with regular variables
print(f(c(1,-1,2,4)))

# run function with second order AD-variables
x <- independent2(c(1,-1,2,4))
y <- f(x)
print(value(y))
print(gradient(y))
print(hessian(y))

}
