% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve,AD_matrix,ANY-method}
\alias{solve,AD_matrix,ANY-method}
\title{Overload of the solve()-function for AD-types}
\usage{
\S4method{solve}{AD_matrix,ANY}(a, b, ...)
}
\description{
Overload of the solve()-function for AD-types
}
