% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve,matrix,AD_matrix-method}
\alias{solve,matrix,AD_matrix-method}
\title{Overload of the solve()-function for AD-types}
\usage{
\S4method{solve}{matrix,AD_matrix}(a, b, ...)
}
\description{
Overload of the solve()-function for AD-types
}
