% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad.R
\name{value}
\alias{value}
\title{Get value from AD type}
\usage{
value(y)
}
\arguments{
\item{x}{An ADtype or AD_matrix type (overloaded also for numeric and matrix types)}
}
\value{
The value of the AD type
}
\description{
Get value from AD type
}
\examples{
# define some function
f <- function(x){ 
M <- matrix(x,2,2)
return(sum(solve(M,c(-1,2))))
}

# run function with regular variables
print(f(c(1,-1,2,4)))

# run function with second order AD-variables
x <- independent2(c(1,-1,2,4))
y <- f(x)
print(value(y))
print(gradient(y))
print(hessian(y))
}
